+++ src/csbm/models/quantized_images.py
@@ -96,11 +96,11 @@ class LatentD3PM(nn.Module):
 
     def markov_sample(self, x: torch.Tensor, t: torch.Tensor, prior: Prior):
         r"""Samples from $p(x_{t-1} | x_{t}, \hat{x_{0}})$, where $\hat{x_{0}} \sim m_{\theta}(\hat{x_{0}} | x_{t})$."""
-        first_step = (t == 1).view((x.shape[0], *[1] * (x.dim() - 1))).to(dtype=x.dtype)
+        first_step = (t == 1).long().view((x.shape[0], *[1] * (x.dim() - 1)))
         
         pred_x_start_logits = self(x, t)
         pred_q_posterior_logits = prior.posterior_logits(pred_x_start_logits, x, t, logits=True)
-        noise = torch.rand_like(pred_q_posterior_logits, dtype=x.dtype)
+        noise = torch.rand_like(pred_q_posterior_logits)
         noise = torch.clamp(noise, min=torch.finfo(noise.dtype).tiny, max=1.)
         gumbel_noise = -torch.log(-torch.log(noise))
         random_samples = torch.argmax(pred_q_posterior_logits + gumbel_noise, dim=-1)
@@ -118,7 +118,7 @@ class LatentD3PM(nn.Module):
     @torch.no_grad()
     def sample(self, x: torch.Tensor, prior: Prior) -> torch.Tensor:
         for t in reversed(range(1, self.num_timesteps + 2)):
-            t = torch.tensor([t] * x.shape[0], device=self.device, dtype=x.dtype)
+            t = torch.tensor([t] * x.shape[0], device=self.device)
             x = self.markov_sample(x, t, prior)
         return x
     
@@ -126,7 +126,7 @@ class LatentD3PM(nn.Module):
     def sample_trajectory(self, x: torch.Tensor, prior: Prior) -> torch.Tensor:
         trajectory = [x]
         for t in reversed(range(1, self.num_timesteps + 2)):
-            t = torch.tensor([t] * x.shape[0], device=self.device, dtype=x.dtype)
+            t = torch.tensor([t] * x.shape[0], device=self.device)
             x = self.markov_sample(x, t, prior)
             trajectory.append(x)
         trajectory = torch.stack(trajectory, dim=0)
